Moves before_invoke and user_action_log to __init__
